requirements.txt
# Core
psycopg2-binary>=2.9.9
streamlit>=1.35.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# Utilities
python-dotenv>=1.0.1
